[feat] separate validator image
